[Refactor] Replace unquantized_gemm monkey-patch with proper custom op dispatch
